feat: Add Dodge Component and System for player dodging mechanics
- Introduced DodgeComponent to manage dodge timing and cooldowns.
- Created DodgeEvent to handle dodge actions.
- Implemented DodgeSystem to process dodge events and manage particle emitters for visual effects.
- Enhanced ParticleEmitterComponent to support new lighting configurations.
- Updated PlayerSystem to trigger dodge events on player input.
- Modified lighting effects to include intensity and attenuation parameters.
- Added EventBus for event management between systems.
- Implemented CollisionResolutionSystem to handle collision events and integrate with dodge mechanics.
- Refactored various components and systems for improved performance and maintainability.

+++ Engine/Particles/ParticleEmitterArc.cs
@@ -1,7 +1,7 @@
 namespace Engine.Particles;
 
-public class ParticleEmitterArc(int minAngle, int maxAngle) : IParticleEmitterShape
+public class ParticleEmitterArc(float minAngle, float maxAngle) : IParticleEmitterShape
 {
-    public int MinAngle { get; set; } = minAngle;
-    public int MaxAngle { get; set; } = maxAngle;
+    public float MinAngle { get; set; } = minAngle;
+    public float MaxAngle { get; set; } = maxAngle;
 }
